Fix exception thrown when renaming a file
Originally, the RenameFile was returning an integer when Dbus was expecting it to return None.  This caused an exception to be thrown.
CreateFileFromTemplate also uses RenameFile and expect it to return an integer, so I moved the RenameFile logic to a separate function,
custom_rename_file which returns an integer so both CreateFileFromTemplate and RenameFile methods can use it